Listen on :PORT
ListenAndServe uses net.Listen (https://golang.org/pkg/net/#Listen),
which "if the host in the address parameter is empty or a literal
unspecified IP address, Listen listens on all available unicast and
anycast IP addresses of the local system."

There may be some systems where "localhost" is not correct.

I also switched to checking the PORT environment variable, which is
standard for where a container should listen for connections. So, while
not neccessary for running locally, it's a good practice to check for
the variable and fall back to a default if needed.

Finally, I updated the startup logging message to include the address
being listened on.